mcp-client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gradio-chat-client"
version = "0.1.0"
description = "Gradio 기반 채팅 애플리케이션"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "개발동생"}
]
dependencies = [
    "gradio==5.29.1",
    "anthropic==0.51.0",
    "python-dotenv==1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[tool.hatch.build.targets.wheel]
packages = ["."] 